DB - models: move get_big_cty as a static method of DXCCEntity
